fix: partially fix 'creating a new project clones the old one'
This bug stil occurs when when creating a project for the first
time (The first clone is intentional although it will keep cloning and
will presist until restarting the program).

The current workaround is to create your first project then instantly
restart the program. After this the bug should be gone forever.

It's not perfect but it's all I can get working right now.