feat(issues): reuse caldav client instances.
To get a list of caldav todos we have to connect the client and discover
the right calendar. This causes at least three requests before we're
finally able to fetch the todos. To mitigate this we cache an
established client connection as well as the discovered calendar
instances - so only the first request for a resource causes the
discovery.