feat(task): keep text on lost focus when adding task
The content of the task input is saved whenever it is nintentionally closed
(by clicking somewhere outside of the task add bar, or because another window
comes to the front). When the task add bar is brought up again, the saved
text is pre-filled and selected (so that it can easily be discarded if not
needed anymore). Previously, the entered text was lost.

Closes #635