fix: avoid nginx error when webdav is missing
By default, nginx refuses to start if the upstream host is not found.
Since the webdav upstream server is optional here, we want nginx to
start even if the webdav upstream server is missing. The trick here is
to put the upstream url in a variable first so that nginx will not check
whether the upstream host exists on start and start anyway.