Fix #4184: Added fractional time parsing when creating a new task
Fixed the issue where you could not insert 1.5h or any decimal value
when inputing time stamp in the quick add method of creating a task. To
test execute project with instructions given in the repository and add a
new task (shift+A) and write the title of a new task such as "task 1.5h"
or "task 1.5h/3.5h". Also added test cases for this fix.