Implementing Repeating Subtasks
This feature enhances the repeating task functionality by allowing
subtasks to be recreated alongside their parent task. Previously,
when a parent task with subtasks was set to repeat, only the parent
would be duplicated on the next scheduled day, leaving the subtasks
behind. With this update, the entire task structure, including all
subtasks, is carried over, ensuring that recurring task groups remain
complete and functional.

This is relevant to the feature request Re-create sub tasks for repeating tasks #427

We took a bit of a different aproach from what was requested, since we felt that
having it in the repeated task menu would be better as it would seperate from
local instances subtasks and repeated subtasks more easily.

Co-authored-by: Tomás Morais <tomas.c.morais@tecnico.ulisboa.pt>