perf: prevent unnecessary change change detections with high-frequency listeners
Listeners schedule change detection, even in zoneless, so those that are
high frequency should be used sparingly. The listeners in this commit do not require
any template refreshes most of the time and should instead use native
`addEventListener` to avoid unnecessary refreshes. While the application
uses `OnPush`, doing a small amount of a lot of times can still have
performance implications.

The `dragover` listener _never_ requires template updates and the resize
listener only requires it when signals update (and the scheduling is
handled by the signal internals)

https://github.com/angular/angular/issues/61839