dialog-confirm: change default to 'OK'
For dialog-confirm modals, change the initially selected / default button from 'cancel' to 'ok' (by switching the tabindex of both). This should make 'OK' the default action when pressing Return on a keyboard, and thus assuming that the user knows what they're doing, and don't need an extre keypress (right arrow) to confirm creating a new tag.