feat(trello): initial working on trello integration.
This is still pretty much WIP; things may break and it still needs to do several thing manually, so I need your feedback

I have tested this and so far it can retrieve the card from trello and added it. Minor visual bugs here and there is seen, so that is one aspect. Automation and ease of task creation should be next.